- working on improved deep copy with interdependency update, so we can at least properly deep copy a network, a graph, a demands, etc. NOT DONE. continue with VirtualNetwork/Centroids containers

+++ src/main/java/org/goplanit/zoning/ConnectoidImpl.java
@@ -236,7 +236,7 @@ public abstract class ConnectoidImpl extends ExternalIdAbleImpl implements Conne
    * {@inheritDoc}
    */
   @Override
-  public long getNumberOfAccessZones() {
+  public int getNumberOfAccessZones() {
     return accessZones.size();
   }
